fix: the sound is creepy

+++ utils/morse_code.py
@@ -61,11 +61,11 @@ MORSE_CODE_DICT = {
 def generate_tone(duration, frequency=700, sample_rate=44100):
     t = np.linspace(0, duration, int(sample_rate * duration), endpoint=False)
     tone = 0.5 * np.sin(2 * np.pi * frequency * t)  # 0.5 cuz otherwise it's loud af
-    return tone
+    return tone.astype(np.int16)
 
 
 def generate_silence(duration, sample_rate=44100):
-    return np.zeros(int(sample_rate * duration))
+    return np.zeros(int(sample_rate * duration), dtype=np.int16)
 
 
 def text_to_morse(text):
@@ -79,7 +79,7 @@ def text_to_morse(text):
 
 
 def morse_to_audio(morse_code, dot_duration=0.1, dash_duration=0.3, sample_rate=44100):
-    audio = np.array([], dtype=np.float32)
+    audio = np.array([], dtype=np.int16)
     for symbol in morse_code:
         if symbol == ".":
             audio = np.concatenate(
